app/models: extract side effects
Currently importing app/models triggers app/pkg/jwt, which causes the
init in app/pkg/env to invoke. While this may be a foregone conclusion
in some contexts, data models should be pure.

This change extracts the jwt model into a new package app/models/jwt,
while forwarding all refrences, except app/models.CreateTenant, through
the old app/pkg/jwt, to reduce noise and because one package is a
better interface than two.
