+++ app/pkg/jwt/jwt.go
@@ -2,6 +2,7 @@ package jwt
 
 import (
 	jwtgo "github.com/dgrijalva/jwt-go"
+	models "github.com/getfider/fider/app/models/jwt"
 	"github.com/getfider/fider/app/pkg/env"
 	"github.com/getfider/fider/app/pkg/errors"
 )
@@ -19,22 +20,10 @@ const (
 )
 
 // FiderClaims represents what goes into JWT tokens
-type FiderClaims struct {
-	UserID    int    `json:"user/id"`
-	UserName  string `json:"user/name"`
-	UserEmail string `json:"user/email"`
-	Origin    string `json:"origin"`
-	Metadata
-}
+type FiderClaims = models.FiderClaims
 
 // OAuthClaims represents what goes into temporary OAuth JWT tokens
-type OAuthClaims struct {
-	OAuthID       string `json:"oauth/id"`
-	OAuthProvider string `json:"oauth/provider"`
-	OAuthName     string `json:"oauth/name"`
-	OAuthEmail    string `json:"oauth/email"`
-	Metadata
-}
+type OAuthClaims = models.OAuthClaims
 
 // Encode creates new JWT token with given claims
 func Encode(claims jwtgo.Claims) (string, error) {
